Separate out the CredentialsSection
