+++ Source/SelfService/Web/integrations/connection/connectionDetails/exposeData/index.tsx
@@ -1,13 +1,14 @@
 // Copyright (c) Dolittle. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
-import React, { useState } from 'react';
+import React from 'react';
 
 import { useSnackbar } from 'notistack';
 
-import { Box, Collapse, FormHelperText, Grid, TextField, Typography } from '@mui/material';
+import { Box, Typography } from '@mui/material';
 
-import { Button, ContentContainer, ContentHeader, ContentSection, IconButton, Link, Switch } from '@dolittle/design-system';
+import { ContentContainer, ContentHeader, ContentSection, IconButton, Link, Switch } from '@dolittle/design-system';
+import { CredentialsSection } from './Credentials/CredentialsSection';
 
 const styles = {
     display: 'flex',
@@ -20,12 +21,10 @@ const styles = {
 
 const restApiUrl = 'https://inspiring-ritchie.dolittle.cloud/erpreadmodels/swagger/index.html';
 const openApiDocumentationUrl = 'https://inspiring-ritchie.dolittle.cloud/erpreadmodels/swagger/v1/swagger.json';
-const credentialToken = 'n$H8rAp3mDJGiR7Adn4@paAzQ7J$cNJSEkzqPDYi';
 
 export const ExposeDataView = () => {
     const { enqueueSnackbar } = useSnackbar();
 
-    const [openCredentials, setOpenCredentials] = useState(false);
 
     const handleRestApiLinkCopy = () => {
         navigator.clipboard.writeText(restApiUrl);
@@ -37,10 +36,6 @@ export const ExposeDataView = () => {
         enqueueSnackbar('OpenAPI documentation copied to clipboard.');
     };
 
-    const handleTokenCopy = () => {
-        navigator.clipboard.writeText(credentialToken);
-        enqueueSnackbar('Token copied to clipboard.');
-    };
 
     return (
         <ContentContainer>
@@ -82,43 +77,7 @@ export const ExposeDataView = () => {
                     />
                 </Box>
             </ContentSection>
-            <ContentSection
-                title='Credentials'
-                headerProps={{ buttons: [{ label: 'Generate New Credentials', variant: 'outlined', onClick: () => setOpenCredentials(true) }] }}
-            >
-                <Collapse in={openCredentials}>
-                    <Grid container spacing={3} sx={{ pb: 5 }}>
-                        <Grid item>
-                            <Typography sx={{ mb: 2 }}>Who or what are these credentials for?</Typography>
-                            <TextField id='credentialsName' label='Name' size='small' />
-                        </Grid>
-
-                        <Grid item>
-                            <Typography sx={{ mb: 2 }}>Credential Token</Typography>
-                            <TextField
-                                id='credentialsToken'
-                                label='Token'
-                                size='small'
-                                type='password'
-                                value={credentialToken}
-                                disabled
-                                sx={{ width: 400 }}
-                            />
-                            <FormHelperText>This bearer token should be used in the request header.</FormHelperText>
-                        </Grid>
-
-                        <Grid container item spacing={3} xs={6} sx={{ alignItems: 'center', mt: 0 }}>
-                            <Grid item>
-                                <Button label='Copy Token' startWithIcon='CopyAllRounded' onClick={handleTokenCopy} />
-                            </Grid>
-
-                            <Grid item>
-                                <Button label='Delete credentials' startWithIcon='DeleteRounded' onClick={() => setOpenCredentials(false)} />
-                            </Grid>
-                        </Grid>
-                    </Grid>
-                </Collapse>
-            </ContentSection>
+            <CredentialsSection />
         </ContentContainer>
     );
 };
